Refactor associated idx
